feat: 모집 정보 및 기술 스텍 매핑 테이블 추가
