+++ src/main/java/durikkiri/project/entity/post/TechnologyStack.java
@@ -2,12 +2,11 @@ package durikkiri.project.entity.post;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -21,9 +20,8 @@ public class TechnologyStack {
     @Column(name = "technology_stack_id")
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "recruitment_info_id")
-    private RecuruitmentInfo recuruitmentInfo;
+    @OneToMany(mappedBy = "technologyStack")
+    private List<RecruitmentTechStack> recruitmentTechStackList;
 
     private String name;
